+++ src/components/Carousel/utils.js
@@ -1,6 +1,6 @@
 import { isInternalURL } from '@plone/volto/helpers';
 
-export function getTeaserImageURL(href, image) {
+export function getTeaserImageURL(href, image, imageType) {
   if (image) {
     if (isInternalURL(image['@id'])) {
       return `${image['@id']}/@@images/image/teaser`;
@@ -8,6 +8,10 @@ export function getTeaserImageURL(href, image) {
       return image['@id'];
     }
   } else {
-    return `${href['@id']}/@@images/preview_image/teaser`;
+    if (imageType === 'image') {
+      return `${href['@id']}/@@images/image/teaser`;
+    } else {
+      return `${href['@id']}/@@images/preview_image/teaser`;
+    }
   }
 }